feat: complete phases 1-3 and align schema with C# project
- Rebuilt database schema to match C# backend (Pascal Case columns,
  same tables: CarsCategories, Admins, Customers, Cars, Rentals, Payments)
- Added Status, CreatedAt, BookingNumber fields across all entities
- Updated all entities with explicit @Column(name=...) mappings
- Fixed type mismatch: Rental dates changed from LocalDate to LocalDateTime
- Added CarService.saveCar() auto-sets CreatedAt on new inserts
- Customer.id changed to Integer to match MySQL INT schema
- JWT authentication wired into SecurityConfig with role-based access
- Admin and Customer login endpoints returning signed JWT tokens
- Global 401/403 JSON error handling (no HTML redirects)
- Updated README with full roadmap, API table, and setup instructions

+++ car-rental/src/main/java/com/nextcar/carrental/controller/CustomerController.java
@@ -26,11 +26,11 @@ public class CustomerController {
 
     // GET /customers/5
     @GetMapping("/{id}")
-    public ResponseEntity<Customer> getCustomerById(@PathVariable Long id) {
+    public ResponseEntity<Customer> getCustomerById(@PathVariable Integer id) {
         return ResponseEntity.ok(customerService.getCustomerById(id));
     }
 
-    // POST /customers/register - 201 Created
+    // POST /customers/register — 201 Created
     @PostMapping("/register")
     public ResponseEntity<Customer> registerCustomer(@Valid @RequestBody CustomerRegistrationDTO dto) {
         Customer created = customerService.registerCustomer(dto);
@@ -40,7 +40,7 @@ public class CustomerController {
     // PUT /customers/5
     @PutMapping("/{id}")
     public ResponseEntity<Customer> updateCustomer(
-            @PathVariable Long id,
+            @PathVariable Integer id,
             @Valid @RequestBody CustomerRegistrationDTO dto) {
         Customer updated = customerService.updateCustomer(id, dto);
         return ResponseEntity.ok(updated);
@@ -48,7 +48,7 @@ public class CustomerController {
 
     // DELETE /customers/5
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCustomer(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteCustomer(@PathVariable Integer id) {
         customerService.deleteCustomer(id);
         return ResponseEntity.noContent().build();
     }
